ENGR00081856: Add SCC2 & RNGC driver support for i.MX37.
Added rng & sahara library files for scc2 and RNGC tests for i.MX37.

+++ rng/fsl_shw_rand.c
@@ -1,6 +1,6 @@
 /*
  * User Space library to access the Security hardware
- * Copyright (C) 2005-2006 written by Freescale Semiconductor
+ * Copyright (C) 2005-2008 written by Freescale Semiconductor
  *
  * This library is free software; you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -17,8 +17,10 @@
  * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301
  * USA
  */
-/*!
- * @file lib/rng/fsl_shw_rand.c
+
+
+/**
+ * @file fsl_shw_rand.c
  *
  * This file implements Random Number Generation functions of the FSL SHW API
  * in USER MODE for talking to a standalone RNGA/RNGC device driver.
@@ -39,89 +41,99 @@
 #include <stdio.h>
 #include <errno.h>
 #include <string.h>
-#endif				/* FSL_DEBUG */
+#endif /* FSL_DEBUG */
 
 #include "shw_driver.h"
 
-extern fsl_shw_return_t validate_uco(fsl_shw_uco_t * uco);
+
+extern fsl_shw_return_t validate_uco(fsl_shw_uco_t *uco);
+
 
 #if defined(FSL_HAVE_RNGA) || defined(FSL_HAVE_RNGC)
 
+
 /* REQ-S2LRD-PINTFC-API-BASIC-RNG-002 */
-fsl_shw_return_t fsl_shw_get_random(fsl_shw_uco_t * user_ctx,
-				    uint32_t length, uint8_t * data)
+fsl_shw_return_t fsl_shw_get_random(
+                                fsl_shw_uco_t* user_ctx,
+                                uint32_t length,
+                                uint8_t* data)
 {
-	fsl_shw_return_t ret = FSL_RETURN_ERROR_S;
+    fsl_shw_return_t ret = FSL_RETURN_ERROR_S;
 
-	/* perform a sanity check / update uco */
-	ret = validate_uco(user_ctx);
-	if (ret == FSL_RETURN_OK_S) {
-		struct get_random_req *req = malloc(sizeof(*req));
+    /* perform a sanity check / update uco */
+    ret = validate_uco(user_ctx);
+    if (ret == FSL_RETURN_OK_S) {
+        struct get_random_req* req = malloc(sizeof(*req));
 
-		if (req == NULL) {
-			ret = FSL_RETURN_NO_RESOURCE_S;
-		} else {
+        if (req == NULL) {
+            ret = FSL_RETURN_NO_RESOURCE_S;
+        } else {
 
-			init_req(&req->hdr, user_ctx);
-			req->size = length;
-			req->random = data;
+            init_req(&req->hdr, user_ctx);
+            req->size = length;
+            req->random = data;
 
-			ret =
-			    send_req(SHW_USER_REQ_GET_RANDOM, &req->hdr,
-				     user_ctx);
-		}
-	}
+            ret = send_req(SHW_USER_REQ_GET_RANDOM, &req->hdr, user_ctx);
+        }
+    }
 
-	return ret;
+    return ret;
 }
 
-fsl_shw_return_t fsl_shw_add_entropy(fsl_shw_uco_t * user_ctx,
-				     uint32_t length, uint8_t * data)
+
+fsl_shw_return_t fsl_shw_add_entropy(
+                                fsl_shw_uco_t* user_ctx,
+                                uint32_t length,
+                                uint8_t* data)
 {
-	fsl_shw_return_t ret = FSL_RETURN_ERROR_S;
+    fsl_shw_return_t ret = FSL_RETURN_ERROR_S;
 
-	/* perform a sanity check on the uco */
-	ret = validate_uco(user_ctx);
-	if (ret == FSL_RETURN_OK_S) {
-		struct add_entropy_req *req = malloc(sizeof(*req));
+    /* perform a sanity check on the uco */
+    ret = validate_uco(user_ctx);
+    if (ret == FSL_RETURN_OK_S) {
+        struct add_entropy_req* req = malloc(sizeof(*req));
 
-		if (req == NULL) {
-			ret = FSL_RETURN_NO_RESOURCE_S;
-		} else {
-			init_req(&req->hdr, user_ctx);
-			req->size = length;
-			req->entropy = data;
+        if (req == NULL) {
+            ret = FSL_RETURN_NO_RESOURCE_S;
+        } else {
+            init_req(&req->hdr, user_ctx);
+            req->size = length;
+            req->entropy = data;
 
-			ret =
-			    send_req(SHW_USER_REQ_ADD_ENTROPY, &req->hdr,
-				     user_ctx);
-		}
-	}
+            ret = send_req(SHW_USER_REQ_ADD_ENTROPY, &req->hdr, user_ctx);
+        }
+    }
 
-	return ret;
+    return ret;
 }
 
-#else				/* no H/W RNG block */
+#else  /* no H/W RNG block */
 
-fsl_shw_return_t fsl_shw_get_random(fsl_shw_uco_t * user_ctx,
-				    uint32_t length, uint8_t * data)
+
+fsl_shw_return_t fsl_shw_get_random(
+                                fsl_shw_uco_t* user_ctx,
+                                uint32_t length,
+                                uint8_t* data)
 {
 
-	(void)user_ctx;
-	(void)length;
-	(void)data;
+    (void)user_ctx;
+    (void)length;
+    (void)data;
 
-	return FSL_RETURN_ERROR_S;
+    return FSL_RETURN_ERROR_S;
 }
 
-fsl_shw_return_t fsl_shw_add_entropy(fsl_shw_uco_t * user_ctx,
-				     uint32_t length, uint32_t * data)
+
+fsl_shw_return_t fsl_shw_add_entropy(
+                                fsl_shw_uco_t* user_ctx,
+                                uint32_t length,
+                                uint8_t* data)
 {
 
-	(void)user_ctx;
-	(void)length;
-	(void)data;
+    (void)user_ctx;
+    (void)length;
+    (void)data;
 
-	return FSL_RETURN_ERROR_S;
+    return FSL_RETURN_ERROR_S;
 }
 #endif
